Fix Admin Controller to get the current user from the JWT instead of the 'g' global variable.some bug fixes on user controller, add @handle_errors

+++ server/controllers/admin_controller.py
@@ -3,8 +3,6 @@ This module provides a controller for the Administrator view
 in the Job-linker application.
 """
 
-from flask import g
-
 from server.exception import UnauthorizedError
 from server.models import storage
 from server.models.application import Application
@@ -29,14 +27,15 @@ class AdminController:
         """
         pass
 
-    def _check_admin(self):
+    def _check_admin(self, curr_user_id):
         """Check if the user is authenticated and is an admin."""
-        if g.user is None:
+        user = storage.get(User, curr_user_id)
+        if user is None:
             raise UnauthorizedError("User is not authenticated")
-        if not g.user.is_admin:
+        if not user.is_admin:
             raise UnauthorizedError("Unauthorized")
 
-    def get_all_users(self):
+    def get_all_users(self, curr_user_id):
         """
         Fetches all users in the system.
 
@@ -48,88 +47,91 @@ class AdminController:
         Returns:
             list: A list of all users.
         """
-        self._check_admin()
+        self._check_admin(curr_user_id)
 
         users = storage.all(User).values()
         return users
 
-    def delete_user(self, user_id):
+    def delete_user(self, target_user_id, curr_user_id):
         """
         Deletes a specific user from the system.
 
         This method should only be accessible by admin users.
 
         Args:
-            user_id (str): The ID of the user to delete.
+            target_user_id (str): The ID of the user to delete.
 
         Raises:
             UnauthorizedError: If the current user is not an admin.
         """
-        self._check_admin()
+        self._check_admin(curr_user_id)
 
-        user = storage.get(User, user_id)
+        user = storage.get(User, target_user_id)
         if not user:
             raise ValueError("User not found.")
+
+        if target_user_id == curr_user_id:
+            raise ValueError("Ask another Admin to Delete you.")
 
         storage.delete(user)
         storage.save()
 
-    def disable_user(self, user_id):
+    def disable_user(self, target_user_id, curr_user_id):
         """
         Disables a specific user account.
 
         This method should only be accessible by admin users.
 
         Args:
-            user_id (str): The ID of the user to disable.
+            target_user_id (str): The ID of the user to disable.
 
         Raises:
             UnauthorizedError: If the current user is not an admin.
         """
-        self._check_admin()
+        self._check_admin(curr_user_id)
 
-        user = storage.get(User, user_id)
+        user = storage.get(User, target_user_id)
         if not user:
             raise ValueError("User not found.")
 
         user.verified = False
         storage.save()
 
-    def enable_user(self, user_id):
+    def enable_user(self, target_user_id, curr_user_id):
         """
         Enables a specific user account.
 
         This method should only be accessible by admin users.
 
         Args:
-            user_id (str): The ID of the user to enable.
+            target_user_id (str): The ID of the user to enable.
 
         Raises:
             UnauthorizedError: If the current user is not an admin.
         """
-        self._check_admin()
+        self._check_admin(curr_user_id)
 
-        user = storage.get(User, user_id)
+        user = storage.get(User, target_user_id)
         if not user:
             raise ValueError("User not found.")
 
         user.verified = True
         storage.save()
 
-    def change_user_role(self, user_id, new_role):
+    def change_user_role(self, target_user_id, new_role, curr_user_id):
         """
         Changes the role of a specific user.
 
         This method should only be accessible by admin users.
 
         Args:
-            user_id (str): The ID of the user.
+            target_user_id (str): The ID of the user.
             new_role (str): The new role for the user.
 
         Raises:
             UnauthorizedError: If the current user is not an admin.
         """
-        self._check_admin()
+        self._check_admin(curr_user_id)
 
         valid_roles = ["candidate", "recruiter"]
         if new_role not in valid_roles:
@@ -138,14 +140,14 @@ class AdminController:
                 f"Role must be one of {valid_roles}."
             )
 
-        user = storage.get(User, user_id)
+        user = storage.get(User, target_user_id)
         if not user:
             raise ValueError("User not found.")
 
         user.role = new_role
         storage.save()
 
-    def get_sys_statistics(self):
+    def get_sys_statistics(self, curr_user_id):
         """
         Fetches system statistics.
 
@@ -158,7 +160,7 @@ class AdminController:
             Dict: A count of all users, recruiters, candidates, jobs,
             applications, languages, skills, majors, and work experiences.
         """
-        self._check_admin()
+        self._check_admin(curr_user_id)
 
         models = {
             "total_users": User,
